add conv2d benchmark
