+++ benchmarks/OpOptimization/CMakeLists.txt
@@ -1,3 +1,3 @@
 add_subdirectory(Conv2dNchwFchw)
 add_subdirectory(MatMul)
-add_subdirectory(Conv2d)
+add_subdirectory(Conv2dOp)
