--- benchmarks/OpOptimization/Conv2d/Main.cpp
@@ -1,32 +1,0 @@
-//===- Main.cpp -----------------------------------------------------------===//
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-//===----------------------------------------------------------------------===//
-//
-// This is the main file of the gemm benchmark.
-//
-//===----------------------------------------------------------------------===//
-
-#include <benchmark/benchmark.h>
-
-void verification();
-
-int main(int argc, char **argv) {
-  // Run benchmark.
-  ::benchmark::Initialize(&argc, argv);
-  ::benchmark::RunSpecifiedBenchmarks();
-  // Run correctness verification.
-  verification();
-  return 0;
-}
